refactor(resource-coordinator): make view strategy explicit
This commit adds an explicit ViewStrategy concept and pipes that
through all aspects of the templating system that deal with custom or
anonymous elements. Logic for determining the default view strategy is
now centralized. The resource coordinator has been patched up to handle
more types of element loading scenarios. Some internal renaming and
duplication was removed in the resource coordinator.