feat(all): support template part replacement
This closes #70. Now developers of custom elements can add a “part”
attribute to any of their internal template controller’s templates.
Then the consumer of their custom element can add a <template> tag in
the content with a “replace-part” attribute set to the same name and
the custom template will override the internal tempalte.