feat(bindable-property): all behaviors now support propertyChanged
This is a minor BREAKING CHANGE which affects only those who
previously used completely dynamic custom attributes. The callback
handler, previously named “dynamicPropertyChanged” and been renamed to
“propertyChanged”. You can now add a “propertyChanged” callback to any
custom element or custom attribute and you will receive all change
notifications for that behavior. This can be done instead of or in
addition to individual property change handlers.