fix(children): child sync can now decorate a property
This is a BREAKING CHANGE. The syncChildren decorator has been renamed
to sync. It is now preferred to use it on a property like this:

@sync(‘tab-item’) items;

Now the single value can specify the CSS selector. This items property
will be populated with child matches of this selector. If there is an
itemsChanged method, it will be called with the items change.

This decorator can still be used on the class as well, or through the
decorators fluent API. In that case it takes an object with name,
changeHandler and selector properties. The changeHandler can still be
determined by convention, but the name and selector are both required
with placed on a class.