feat(view-compiler): enable surrogate behaviors, bindings and attributes
For custom elements, in their view, you can now add attributes,
bindings and custom attribute behaviors directly to the template
element itself. The template element serves as a surrogate for the
actual custom html element that will be created. When it is created all
attributes, bindings and custom attribute behaviors will be copied over
to the custom element. Their binding context will be the same as the
custom element’s view, namely the custom element class instance. This
simplifies custom elements that need to dynamically add classes to
themselves or update aria attributes. This also makes it easy to apply
cross-cutting concerns directly to custom element by simply factoring
them into custom attributes which are then placed on the template tag
itself.

Closes #61