feat(html-behavior): enable extensible content processing
This deprecates the @skipContentProcessing decorator in favor of using
the @processContent decorator. To achieve the same affect, you would
use @processContent(false). But now, you can supply a function to the
decorator like so @processContent(myCustomProcessor) and that function
will be invoke when an instance of the custom element is compiled. This
gives the ability to the developer to process content during the
compilation phase. The signature of the custom processing function is
function(compiler:ViewCompiler, resources:ResourceRegistry, node:Node,
instruction:Object):boolean Return true to have the compiler perform
its own processing on the content after the custom processing, false
otherwise. The calling context for this function is the HTMLBehavior
instance itself. Thus the custom processor has access to all the
instance’s metadata.