feat(all): major view refactor
Views now no longer have start/and comments and instead track their
nodes internally. Slots now insert after the anchor because there’s no
way to find the beginning of a view without a marker, but we can find
the end. Everything seems to be working as well as before.