fix(templating-engine): enhanced views should attach
This ensures that attached is called on the enhanced view before it is
returned. Additionally, this commit fixes the firstChild, lastChild and
fragment properties of an enhanced view so that now the view can be
removed from the DOM and even re-added through a ViewSlot later.

Fixes https://github.com/aurelia/framework/issues/724