fix(view-compiler): Fix reference to next sibling node
The _compileElement function returns the nextSibling of the compiled node. If you were to manipulate the element content however in such a way that the element was no longer a child of its original parent, the original next sibling would never be reached. This change keeps and uses a reference to the original next sibling element.